requirements.txt
alembic==1.8.1
aniso8601==9.0.1
bcrypt==3.2.2
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
mongoengine==0.24.2
packaging==21.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.4.0
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1